+++ src/main/java/org/hosh/runtime/FileSystemCompleter.java
@@ -58,6 +58,11 @@ public class FileSystemCompleter implements Completer {
 			listCandidates(path, p -> p, candidates);
 		} else if (path.isAbsolute()) {
 			listCandidates(parent(path), Path::toAbsolutePath, candidates);
+		} else if (path.getParent() != null) {
+			Path resolved = state.getCwd().resolve(path.getParent());
+			int from = state.getCwd().getNameCount();
+			int to = resolved.getNameCount() + 1;
+			listCandidates(resolved, p -> p.subpath(from, to), candidates);
 		} else {
 			listCandidates(state.getCwd(), Path::getFileName, candidates);
 		}