+++ CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.28] - 2019-03-29
 - new command 'regex', basically named groups regexp to record
 - new command 'network', to disply all network interfaces in this system
+- introducing annotation-based help system
 - improving autocomplete of partially specified directories
-
 
 ## [0.0.27] - 2019-03-17
 - new system commands: